+++ packages/shader-composer-toybox/src/tools/index.ts
@@ -1,2 +1,3 @@
 export * from "./Grid2D"
 export * from "./ModifyVertex"
+export * from "./SceneDepthTexture"
